+++ src/ICP_hackathon_frontend/src/App.jsx
@@ -1,31 +1,25 @@
-import { useState } from 'react';
-import { ICP_hackathon_backend } from 'declarations/ICP_hackathon_backend';
+import "./index.scss";
+import Home from "./Components/Home";
+import About from "./Components/About";
+import Work from "./Components/Work";
+import Testimonial from "./Components/Testimonial";
+import Contact from "./Components/Contact";
+import Footer from "./Components/Footer";
+// import { useState } from 'react';
+// import { ICP_hackathon_backend } from 'declarations/ICP_hackathon_backend';
 
 function App() {
-  const [greeting, setGreeting] = useState('');
-
-  function handleSubmit(event) {
-    event.preventDefault();
-    const name = event.target.elements.name.value;
-    ICP_hackathon_backend.greet(name).then((greeting) => {
-      setGreeting(greeting);
-    });
-    return false;
-  }
-
   return (
-    <main>
-      <img src="/logo2.svg" alt="DFINITY logo" />
-      <br />
-      <br />
-      <form action="#" onSubmit={handleSubmit}>
-        <label htmlFor="name">Enter your name: &nbsp;</label>
-        <input id="name" alt="Name" type="text" />
-        <button type="submit">Click Me!</button>
-      </form>
-      <section id="greeting">{greeting}</section>
-    </main>
+    <div className="App">
+      <Home />
+      <About />
+      <Work />
+      <Testimonial />
+      <Contact />
+      <Footer />
+    </div>
   );
 }
 
 export default App;
+
