+++ src/ICP_hackathon_frontend/src/Components/About.jsx
@@ -1,0 +1,39 @@
+import React from "react";
+import AboutBackground from "../Assets/about-background.png";
+import AboutBackgroundImage from "../Assets/about-background-image.png";
+import { BsFillPlayCircleFill } from "react-icons/bs";
+
+const About = () => {
+  return (
+    <div className="about-section-container">
+      <div className="about-background-image-container">
+        <img src={AboutBackground} alt="" />
+      </div>
+      <div className="about-section-image-container">
+        <img className='all-images' src={AboutBackgroundImage} alt="" />
+      </div>
+      <div className="about-section-text-container">
+        <p className="primary-subheading">About</p>
+        <h1 className="primary-heading">
+          "Event Experiences Are Key to a Balanced Life"
+        </h1>
+        <p className="primary-text">
+          Lorem ipsum dolor sit amet consectetur. Non tincidunt magna non et
+          elit. Dolor turpis molestie dui magnis facilisis at fringilla quam.
+        </p>
+        <p className="primary-text">
+          Non tincidunt magna non et elit. Dolor turpis molestie dui magnis
+          facilisis at fringilla quam.
+        </p>
+        <div className="about-buttons-container">
+          <button className="secondary-button">Learn More</button>
+          <button className="watch-video-button">
+            <BsFillPlayCircleFill /> Watch Video
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default About;
